+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Versions
 
+## x.x.x
+* Gracefully handle inter and rewarded ad calls when current `Activity` is null.
 ## 5.0.1
 * Depend on Android SDK 11.8.2 and iOS SDK 11.8.2.
 ## 5.0.0